apps/api/pyproject.toml
[project]
name = "investara-api"
version = "0.1.0"
description = "FastAPI backend for Investara"
requires-python = ">=3.11,<3.15"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.6.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.0",
  "sqlalchemy>=2.0.0",
  "openai>=1.60.0",
  "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "ruff>=0.8.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

